Move resetting code tests to the change-pin file for consistency
