+++ tests/verify.rs
@@ -1,7 +1,6 @@
 // Copyright (C) 2022 Nitrokey GmbH
 // SPDX-License-Identifier: LGPL-3.0-only
 
-use openpgp_card::StatusBytes;
 use test_log::test;
 
 mod card;
@@ -113,8 +112,6 @@ fn verify() {
             tx.reset_retry_counter_pw1(b"new code", None).unwrap();
             assert_checks!(tx, Some(3), Some(3), None);
             tx.verify_pw1_user(b"new code").unwrap();
-            tx.set_resetting_code(&[0; 127]).unwrap();
-            tx.set_resetting_code(&[0; 128]).unwrap_err();
         });
         card.reset();
         card.with_tx(|mut tx| {
@@ -122,52 +119,6 @@ fn verify() {
             assert!(tx.verify_pw1_user(b"bad code").is_err());
             assert!(tx.verify_pw1_user(b"bad code").is_err());
             assert_checks!(tx, Some(0), Some(0), Some(3));
-            tx.reset_retry_counter_pw1(b"123456", Some(&[0; 127]))
-                .unwrap();
-            assert_checks!(tx, Some(3), Some(3), Some(3));
-            tx.verify_pw1_user(b"123456").unwrap();
-        });
-        card.reset();
-        card.with_tx(|mut tx| {
-            assert!(tx.reset_retry_counter_pw1(b"new code", None).is_err());
-            assert!(tx.verify_pw1_user(b"new code").is_err());
-            assert_eq!(
-                error_to_retries(tx.reset_retry_counter_pw1(b"new code", Some(b"12345678"))),
-                Some(2)
-            );
-            let short_reset = tx.reset_retry_counter_pw1(b"short", Some(&[0; 127]));
-            assert!(
-                matches!(
-                    short_reset,
-                    Err(openpgp_card::Error::CardStatus(
-                        StatusBytes::IncorrectParametersCommandDataField
-                    ))
-                ),
-                "Got: {short_reset:?}"
-            );
-            assert_eq!(
-                error_to_retries(tx.reset_retry_counter_pw1(b"new code", Some(b"12345678"))),
-                Some(2)
-            );
-            assert!(tx.verify_pw1_user(b"new code").is_err());
-            assert_eq!(
-                error_to_retries(tx.reset_retry_counter_pw1(b"new code", Some(b"12345678"))),
-                Some(1)
-            );
-            assert!(tx.verify_pw1_user(b"new code").is_err());
-            assert_eq!(
-                error_to_retries(tx.reset_retry_counter_pw1(b"new code", Some(b"12345678"))),
-                Some(0)
-            );
-            assert!(tx.verify_pw1_user(b"new code").is_err());
-            assert_eq!(
-                error_to_retries(tx.reset_retry_counter_pw1(b"new code", Some(b"12345678"))),
-                Some(0)
-            );
-            assert!(tx.verify_pw1_user(b"new code").is_err());
-            tx.reset_retry_counter_pw1(b"123456", Some(b"1234567890"))
-                .unwrap_err();
-            assert!(tx.verify_pw1_user(b"new code").is_err());
         });
     })
 }
